requirements.txt
# Backpack-Lighter Arbitrage Bot Dependencies

# Core dependencies
python-dotenv>=1.0.0
websockets>=12.0
requests>=2.31.0
tenacity>=8.2.0

# Cryptography for Backpack auth
cryptography>=41.0.0

# Backpack SDK
bpx-py>=0.1.0

# Lighter SDK
lighter-sdk>=1.0.0

# Decimal precision
typing-extensions>=4.8.0

